% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot_response}
\alias{plot_response}
\alias{plot_process_mix}
\alias{plot_dashboard}
\title{diagnostic plots}
\usage{
plot_response(D)

plot_process_mix(D)

plot_dashboard(D)
}
\arguments{
\item{D}{uncanny sim data}
}
\value{
gg
}
\description{
diagnostic plots
}
