% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_explorer.R
\docType{data}
\name{poly_explorer}
\alias{poly_explorer}
\alias{poly_explorer_ui}
\alias{poly_explorer_server}
\title{Interactive polynomial explorer}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
poly_explorer()

poly_explorer_ui

poly_explorer_server(input, output)
}
\value{
interactive shiny app
}
\description{
Interactive polynomial explorer
}
\keyword{datasets}
