% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{shoulder}
\alias{shoulder}
\alias{shoulder.numeric}
\alias{shoulder.matrix}
\alias{shoulder.data.frame}
\title{Derive the local maximum of a third degree polynomial}
\usage{
shoulder(coef, ...)

\method{shoulder}{numeric}(coef)

\method{shoulder}{matrix}(coef, ...)

\method{shoulder}{data.frame}(coef)
}
\arguments{
\item{coef}{polynomial coefficients}

\item{...}{}
}
\value{
position of the local maximum, NA if it does not exist
}
\description{
Takes four polynomial parameters as input and derives the
local minimum if it exists.
}
\examples{

shoulder(c(-1, -4, 3, 1))
}
