% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{trough}
\alias{trough}
\alias{trough.numeric}
\alias{trough.matrix}
\alias{trough.data.frame}
\title{Derive the local minimum of a third degree polynomial}
\usage{
trough(coef, ...)

\method{trough}{numeric}(coef = c(-0.2, -0.5, 0.2, 0.7))

\method{trough}{matrix}(coef, ...)

\method{trough}{data.frame}(coef)
}
\arguments{
\item{coef}{polynomial coefficients}

\item{...}{}
}
\value{
position of the local minimum, NA if it does not exist
}
\description{
Takes four polynomial parameters as input and derives the local minimum
if it exists.
}
\examples{

trough(c(-1, -4, 3, 1))
}
