% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cat_conflict.R
\docType{data}
\name{uv_explorer}
\alias{uv_explorer}
\alias{uv_explorer_ui}
\alias{uv_explorer_server}
\title{Interactive uncanny valley explorer}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
uv_explorer()

uv_explorer_ui

uv_explorer_server(input, output)
}
\value{
interactive shiny app
}
\description{
Interactive uncanny valley explorer
}
\keyword{datasets}
